Do not rotate the log on shutdown if the kernel is unhealthy
NSDS always bumps the log version in the neo store on shutdown and in
such way it behaves differently wrt LogRotationImpl that avoids to
rotate logs in case the kernel is not healthy. This fix tries to unify
the log rotation behavior by avoiding to bump the log version when
shutting down NSDS if the kernel is unhealthy.

Notice that we don't throw exception from NSDS when shutting down
since we assume that the shutdown is always possible without causing
exceptions.