Useful names for generated tests in IndexTxStateLookupTest
We were previously calling toString() on arrays, which was not
helpful. JUnit doesn't give us a sophisticated hook to generate the test
names, just parameter substitution with toString(), so we're laboriously
creating our own names by hand for each case.

This is particularly useful for TeamCity because the fixed names allow
us to see history for individual tests.