Reverts behaviour change of 9a868b869213b755b143152dc9abdb9310c475df
in which a change was made that made removing relationship from legacy
index throw EntityNotFoundException if not found. This reverts back to
behaviour of just skipping that removal.

Problem was that locks in some scenarios were held indefinitely,
seemingly, which caused some problems in clusters using legacy
relationship indexes.

This commit doesn't contain a unit test because the reason why locks were
indefinitely held is unknown, but this commit beings behaviour back to what
it has always been like and has been manually verified with ha-robustness.