Make dbms.querylog.filename be null by default
The idea is to force the user to specify an absolute path when using
embedded and to not log queries when the path is not specified.  This
solves the issue that queries have been logged in the store directory
and that behaviour was confusing for end users.

The null default value is overridden when using server and the queries
are logged by default into 'data/log/queries.log'.