Iterative qsort and more informed pivot selection in ParallelSort
Problem being that the qsort algorithm, which is implemented with
recursion, may be extremely unlucky and hit StackOverflowError before
reaching the leaves to start sorting.

Pivot base is (and was) selected at random. In addition to that this
commit adds so that 10 surrounding items are looked at and the mean
of those selected as the pivot to use. This should result in a better
worst-case.

The qsort is changed from being implemented using recursion to using
an iterative approach with a stack to push partitions instead, this to
remove chance of StackOverflowError being thrown for unlucky cases.

When reading this commit it may look like much have chanced. That's
because a couple of methods have been moved into SortWorker where they
belong. The important change is the added qsort and removed
recursiveQsort and of course the added informedPivot method.