Implemented the cypher parts of mandatory property constraints
Implementing the necessary cypher infrastructure for creating and dropping mandatory
property constraints. Note that nothing is yet implemented on the kernel side so at the moment
using the feature will end up in a `NotImplementedError`. Also note that syntax is not yet final,
however the implementation is not expected to change much due to syntax changes.