Improve kernel panic when applying transactions to the store
Before the transaction commit process was in charge to raise a kernel
panic in case of failure in transaction application and since the
commit process is different between single and HA we had to duplicate
that logic in several places: TransactionRepresentationCommitProcess
and TransactionCommittingResponseUnpacker.

Now the responsibility for kernel panic is moved down into
TransactionRepresentationStoreApplier and IndexUpdatesValidator.
Moreover such a logic is not spread around multiple modules but it
lives only inside the kernel.