Remove server startup timeout
The startup block for the server is no longer wrapped in a timeout. This
startup timout has been severely watered down - upgrades and recovery happen
outside it, and HA has it disabled entirely. As such, it does not guard against
any known blocking startup actions, and remains only for historical reasons.

This commit removes it entirely, which simplifies the code.

We also remove all preflight tasks other than one. The upgrade and recovery
ones because they no longer need to happen outside the timeout, and the
config validating one becuase it was not used and is covered by the new Config
system usage in server instead