Flush the same log channel as the one written to
Since forcing the log channel was not synchronized with the preceding
writing of the buffer to the channel, it was possible that the force of
the channel ended up forcing the next channel. In order to make it
easier to reason about the forcing of the channel we introduce a
Flushable object that we obtain under said sycnronization but flush
outside of it. This gives us the best of both worlds, shorter
synchronized regions, but the ability to reason about what we force.

Note that since the previous API called this method flush() but the new
API uses the existing java.io.Flushable interface, which defines a
method called flush() this description uses these two words almost
interchangeably, but uses force to refer to previous behaviour and
flush when talking about new behaviour. The implementation of flush()
is to call the same underlying force-method as the previous force()
method called, but binds directly to that object instead of through
a volatile lookup.