Adds a thorough test for transaction event data
also makes DatabaseRule implement GraphDatabaseAPI so that
dbRule.getGraphDatabaseService()/getGraphDatabaseAPI() gets deprecated,
since a DatabaseRule can be used as a
GraphDatabaseAPI/GraphDatabaseService directly.
Also changes the default behaviour so that the db is started as part of
the before section of the rule instead of on the first call to
getGDS()/getGDAPI() to behave more like rules usually do. Although there's
the option of having the old behaviour to start lazily. This is done by
calling startLazily() on the DatabaseRule, f.ex:

DatabaseRule db = new EmbeddedDatabaseRule( getClass() ).startLazily()

although very few tests need that functionality.