Removes BatchingPageCache
in favor of plain MuninnPageCache. This to reduce code complexity and the
burden to maintain two implementations of PageCache.

Observed throughput is slightly lower in MuninnPageCache due to the
shortcuts taken in the BatchingPageCache for the specific use case of the
batch importer, but that matters quite little since page cache I/O wasn't
and doesn't seem to have a likeliness to be the bottleneck.

In short the batch import performance isn't affected by this.