Fixes input id iterator stopping on null values
During id collision resolving phase in batch importer, preparing the
IdMapper, the passed in ID iterator would have a flaw in that it stopped
its iteration on the first null ID encountered. This would mess up the
collision resolving in that not all IDs marked as collisions would be
added to the list to be sorted, although the sorting algorithm thought it
was and so there would potentially be some data entries in the end that
wouldn't have been initialized and eventially break the sorting algo.