Make FOREACH cypher queries return empty result
Previously result of `FOREACH( n in range( 0, 1 ) | CREATE (p:Person) )` query
would be Iterator(Map()) which is incorrect.
Empty iterator is the correct and expected result.

This was so because EmptyResultPipe was not planned on top of foreach and
dummy empty execution context, created by SingleRowPipe, was passed
all the way up to the user.

Commit fixes the problem by forcing EmptyResultPipe on top of any pipe that
has nothing to return. To avoid infinite planning of EmptyResultPipe it's
presence is checked in EmptyResultBuilder#canWorkWith.