Fix a bug where vectored flushing can race with eviction
Previously we iterated the translation table and added all pages to our IO vector, that we both loaded and dirty.
However, this is not a sufficiently strict criteria, beucase in between reading the page out of the translation table and locking it, it can be evicted and faulted into another page.
If this happened, we could end up adding a dirty page *from another file* to our IO vector, thus mixing the contents of the file.

To fix this, I have changed the check for whether the page is loaded or not, to a check for whether the page is bound to the expected file and file page id.
The flushing code then computes what file page id we should expect to see, from the location of the page in the translation table.
This also means that we won't be fooled if the page is faulted into the same file but at a different location.

I have added no tests for this, because the data race is too narrow to hit with a targeted test. This bug was found through code inspection.