Fix a pread bug in the EphemeralFSA
The bug was that a positioned read of an empty file did not correctly recognise that we were reading beyond the end of the file.
So, instead of returning -1 as the EOF signal, it would return the negative of the requested file position, which is complete non-sense.
The fix is that we now check for any figure of available bytes that is less than or equal to zero, and return the EOF signal for that.
The old code required that the number of available bytes ended up as exactly zero, which of course is quite rare, since most requested
file positiones are not exactly zero.