Fix an integer overflow bug in the expected max memory allocation
The MuninnPageCache constructor was calling the MemoryManager constructor, giving it an estimate of how much memory we might need.
This estimate was computed by multiplying the maxPages and the cachePageSize values together.
The problem was that both of these variables were 32 bit signed integers, and thus could easily overflow with just a moderate page cache memory allocation.
The effect of this bug, is that the MemoryManager would get a really bad estimate of how much memory would be needed, and thus end up producing many more Slab objects than expected.
In the worst case, this could lead to an OutOfMemoryError: Heap space, when the increased number of Slab objects put too much memory pressure on the Java heap.

The fix is to first cast one the variables to a `long` before doing the multiplication.