Fix vectored IO
The vectored read and write methods did not properly handle the cases where the corresponding methods on the StoreChannel processed fewer bytes than what was asked for.
This has been fixed with retry loops, similar to how the positioned read and write methods do.
This ends up working quite nicely because Java NIO makes sure to update the file and buffer positions.
Testing was a pain, because we had to inject an adversarial version of FileDispatcher into the underlying sun.nio.ch.FileChannelImpl.
To do this, we have to put the AccessibleFileDispatcher and DelegateFileDispatcher on the boot classpath, which in turn means the JVM has to be started with some special parameters.
A custom JUnit test runner, the BootClassPathRunner, was implemented to support this.
The BootClassPathRunner runs the unit tests in a separate process, where the boot classpath has been suitably mangled.
Using this new runner, we now test both the PageCache as a whole, and the SingleFilePageSwapper, using the hacked up FileChannelImpls.