Deadlock on db shutdown caused by NeoStoreDataSource and CheckPointScheduler waiting for each other.
Before db shutdown we will try to shutdown all the life components first.
Life.shutdown was invoked while holding logFile lock. That causing deadlock between NeoStoreDataSource
that holding it and life component - CheckPointScheduler that on stop will wait for currently running checkpoint to finish
if any. In that case checkpointer will be blocked forever because it needs to lock logFile to operate.
Also fix incorrect handling of stop flag in checkpointer job.