StorePropertyCursor made non-lazy
This is done so to guarantee consistent view of a single property record while
loading of a property chain. StorePropertyCursor is now only responsible for
reading of property record header (previousPropertyId and nextPropertyId).
Introduced new StorePropertyPayloadCursor that provides a view of the
property blocks. Both property record header and blocks are initialized
eagerly and using the same version of the page (guaranteed by reading inside
single PageCursor should-retry loop).

Lazy loading of blocks was not correct because it did not ensure all blocks
were read from the same version of the page. Under a heavy concurrent load
this resulted in reading of stray propertyKeyId from property blocks.