Brings in the legacy consistency checker as a separate component
which will live in community/consistency-check-legacy until we're
confident that the new and big changes are stable and works in all
scenarios. The legacy checker is a complete copy of the code base as it
was in 2.3 HEAD right before this commit, and as such can be deleted
without affecting the new checker when that time comes.

The legacy checker will be default for a little while longer, but can be
run if --experimental is supplied.

Up to date with commit be67faa2c0942071705b11828ca0d9753be67d36