Remove trailers from store files since they are redundant
They are _not_ used to determine recovery and they are _not_ useful
for rebuilding id generators since ".id" files have their own way to
check if they were cleanly closed or not.

This commit removes the trailers from the stores.

Note: the counts store relies on the trailer to determine if the
latest version was completely flushed on disk or not.  In order to
preserve this check now we write as trailer the "format specifier"
which can be found in the header as well.

For future reference: Trailers are used before migration to check that
the stores are closed cleanly, since we have trailers in older store
previous 2.3, this is fine.  For post and 2.3 stores the best way of
implementing this would be to check that the latest entry in the
transaction log is a checkpoint.