Fewer calls to NeoStores#getXXXStore()
Instead getting the required store(s) ones and use directly. This is a
good practive, for avoiding that additional method call getting
the store for every interaction.

Before this commit there was additional synchronization on each getter
since stores were openened lazily on first access. This was particularly
noticable in hot execution paths. This is still supported, but the default
behaviour is to eagerly open all stores and if choosing that option
removing the synchronization overhead.

There are still some places where it's very convenient to not pull out
each store, since all stores are used, f.ex. when applying transactions
onto the store. Those are left intact.