StoreCopyServer checkpoint before full backup
Instead of flushing store files on disk and cause a race with the
check pointer, it will ask the check pointer to try check point for
making sure we have flushed store files on disk, and after that it
will start streaming transactions starting from the transaction used
in the check point.

This should solve a failure caused by the race mentioned above where
we were trying to rotate the counts store while it was in a rotation
state.