Fix test isolation failure in PageCacheTest.mustNotLoseUpdates
The test uses the RepeatRule to run multiple times, and was using a file with a fixed name.
It was relying on the ephemeral file system being recreated every run, which is not neccesarily the case for the tests that uses the real file system.
This is fixed by explicitly zeroing out the file pages before we use them.