NeoStore last committed tx will always return the highest committed tx
Before this change we were returning the highest gap free tx from the
OutOfOrderSequence, this was causing problem in a cluster
configuration when a slave should pull txs to become up-to-date wrt to
the master when acquiring locks or after having committed on the
master.  Indedd slaves were pulling not enough transaction and were
basing subsequence changes on out of date records causing
inconsistencies.