TxPushStrategyConfigIT#twoRoundRobin asserts from a user POV
instead of nitty-gritty, down-to-the-last-bit assertions about exact txId
on each slave for each transaction, the assertions in this test have been
rewritten with how user would percieve the pushing. So now a couple of
transactions are performed and the test then verifies that there's at most
a gap of 1 between all members. Hopefully this will remove sporadic
failures.