Ensure forcing is done after recovery completes
Recover forced everything after the transactions had been replayed, this
was not late enough though, since the index recovery is not completed at
this stage. If we instead force everything in the start phase of startup
index recovery will have completed, and forcing will cause the index to
be refreshed to the recovered state. When forcing before the indexes had
recovered, the index accessor would keep stale unrecovered state,
leading to inconsistencies at runtime.