Awaits uniqueness constraint indexes to fully populate during recovery
before returning from the GDS constructor. Previous there was this
mismatch where a newly created uniqueness constraint, after a crash, would
have to be rebuilt from scratch. This is done after the store has
recovered, so that's fine. However the usualt semantics of creating a
uniqneness constraint is that the transaction creating it waits for it to
be fully populated before returning. This added waiting mimics that
behaviour for recovery. Without it recovery could be considered completed
while uniquness constraint indexes were still being populated in the
background.