Revert halfway done work on Identity.
This reverts the partial support for an Identity type that the Bolt protocol
specifies. The reason for this being that it is not compliant with the
proposed Identity type, and that the work on Identity so far has
highlighted the complexity of it.

Eg: Identity is a big epic, this is not in the right direction of that epic,
it is better that the protocol is clean and releasable than that it has
partial broken support for Identity.

Identity is expected to return, but as a larger chunk of work.