Introduce FeatureToggles
FeatureToggles is a helper class that makes it easer to use java System
properties to control certain features. This is done by providing
methods for boolean flag values (with a default value), integer
parameters, and enum flags. All feature toggle System parameter names
are fully qualified with a class name.
The intended use case is for the FeatureToggles methods to be invoked
from a static context and the value stored in a static final field.
The default value is used if the parameter is not set or if the
parameter has an unrecognized value.