Improved how current cluster member role is determined
ClusterMembers now use HighAvailabilityMemberStateMachine to determine role of
the current instance. Since state machine is able to deal with out of order
messages it knows better what the actual role is. State of other cluster members
is tracked by ObservedClusterMembers that uses cluster events to figure out what
roles, uris, etc. There is no guarantee that information in ObservedClusterMembers
is fully up-to-date because of possible events reordering.

Signed-off-by: @MishaDemianenko