Move rrdb files in its own directory
This change is needed since the old default directory was the store
directory, but this was causing issues when branching data, for
example we failed to delete old branched data due to rrd files still
in use.  Moving the rrd files out of that directory will solve this
problem.  The new default directory will be 'data/rrd'.

This adds also a check for the availability of database before
allowing the rrd sampler to sample nodes/relationships/properties.
This should minimize the sampling failures when the db is down due to
cluster reconfigurations.