Fixes relationship cursor issue for dense node w/o relationships
Nodes can be marked as dense when getting new relationships, although
never marked as sparse even after deleted below dense node threshold.
So there can be dense nodes w/o any relationships.

These was an assumption in StoreNodeRelationshipCursor that just because a
node was dense it had to have at least one relationship, and so it went
ahead and loaded a record regardless of the first relationship id, which
could be -1. This commit fixes that.

Closes #5691